refactor: 중복되는 백로그 상수를 상수파일로 관리하도록 수정
