+++ src/components/backlog/StoryFormModal.tsx
@@ -10,6 +10,7 @@ import type {
   CreateStoryRequest,
 } from '@/types/backlog';
 import type { MemberInfo } from '@/types/api';
+import { STATUS_OPTIONS, PRIORITY_OPTIONS, ISSUE_TYPE_OPTIONS } from '@/constants/backlog';
 import AssigneeSelect from './AssigneeSelect';
 
 interface StoryFormModalProps {
@@ -23,25 +24,6 @@ interface StoryFormModalProps {
   onClose: () => void;
   onManageEpics?: () => void;
 }
-
-const STATUS_OPTIONS: { value: StoryStatus; label: string }[] = [
-  { value: 'OPEN', label: '할 일' },
-  { value: 'IN_PROGRESS', label: '진행 중' },
-  { value: 'DONE', label: '완료' },
-  { value: 'CLOSED', label: '종료' },
-];
-
-const PRIORITY_OPTIONS: { value: Priority; label: string }[] = [
-  { value: 'LOW', label: '낮음' },
-  { value: 'MEDIUM', label: '보통' },
-  { value: 'HIGH', label: '높음' },
-  { value: 'URGENT', label: '긴급' },
-];
-
-const ISSUE_TYPE_OPTIONS: { value: IssueType; label: string }[] = [
-  { value: 'FE', label: 'FE' },
-  { value: 'BE', label: 'BE' },
-];
 
 export default function StoryFormModal({
   mode,
